Add schema folder and move all types there to it
