+++ app/vision_automl/schemas/ml_tasks.py
@@ -1,11 +1,7 @@
 from pathlib import Path
-from typing import Annotated, Literal, Union
+from typing import Literal, Annotated, Union
 
 from pydantic import BaseModel, Field
-
-# ---------------------------------------------------------------------------
-# Base models
-# ---------------------------------------------------------------------------
 
 
 class ImageTask(BaseModel):
@@ -27,11 +23,6 @@ class TextTask(BaseModel):
 
     class Config:
         arbitrary_types_allowed = True
-
-
-# ---------------------------------------------------------------------------
-# Image tasks
-# ---------------------------------------------------------------------------
 
 
 class ImageClassificationTask(ImageTask):
@@ -78,11 +69,6 @@ class KeypointDetectionTask(ImageTask):
     task_type: Literal["keypoint_detection"] = "keypoint_detection"
 
 
-# ---------------------------------------------------------------------------
-# Audio tasks
-# ---------------------------------------------------------------------------
-
-
 class AudioClassificationTask(BaseModel):
     """Configuration for audio classification tasks.
 
@@ -96,11 +82,6 @@ class AudioClassificationTask(BaseModel):
 
     class Config:
         arbitrary_types_allowed = True
-
-
-# ---------------------------------------------------------------------------
-# Text tasks
-# ---------------------------------------------------------------------------
 
 
 class SequenceClassificationTask(TextTask):
@@ -149,11 +130,6 @@ class MaskedLMTask(TextTask):
     task_type: Literal["masked_lm"] = "masked_lm"
 
 
-# ---------------------------------------------------------------------------
-# Existing tasks kept for backward compatibility
-# ---------------------------------------------------------------------------
-
-
 class ImageMultiLabelClassificationTask(ImageTask):
     """Configuration for multi-label image classification tasks."""
 
@@ -167,10 +143,6 @@ class ImageRegressionTask(ImageTask):
     task_type: str = "image_regression"
     label_format: Literal["csv"] = "csv"  # regression needs exact values
 
-
-# ---------------------------------------------------------------------------
-# Supported task type slug set (mirrors ml_engine/configs)
-# ---------------------------------------------------------------------------
 
 SUPPORTED_VISION_TASK_TYPES: frozenset[str] = frozenset(
     {
@@ -187,8 +159,6 @@ SUPPORTED_VISION_TASK_TYPES: frozenset[str] = frozenset(
         "masked_lm",
     }
 )
-
-# Discriminated union of all first-class task schemas
 VisionTask = Annotated[
     Union[
         ImageClassificationTask,
